requirements.txt
APScheduler~=3.10.4
flask~=3.0.3
PyYAML~=6.0.2